+++ src/features/assetsDecorator.ts
@@ -3,8 +3,8 @@ import * as fs from 'fs';
 import { resolveGUID } from './guidUtilsProvider';
 import * as utils from '../other/utils';
 import * as path from 'path';
-import { ASSETS_FILENAME_PATTERN, uniqueAssetName, assetNameWithOrigin } from '../other/assetsXml';
-import * as minimatch from 'minimatch';
+import { assetNameWithOrigin } from '../other/assetsXml';
+import * as editorFormats from '../editor/formats';
 
 import { clearDiagnostics, diagnostics, refreshDiagnostics } from './assetsActionProvider';
 
@@ -16,21 +16,10 @@ export function activate(context: vscode.ExtensionContext) {
 
   let timeout: NodeJS.Timer | undefined = undefined;
 
-  // create a decorator type that we use to decorate small numbers
   const guidDecorationType = vscode.window.createTextEditorDecorationType({});
-  const modOpDecorationType = vscode.window.createTextEditorDecorationType({});
   const assetDecorationType = vscode.window.createTextEditorDecorationType({});
 
   let activeEditor = vscode.window.activeTextEditor;
-
-  function shortGuidDecorator(guid: string) {
-    const resolved = resolveGUID(guid);
-    if (!resolved) {
-      return '';
-    }
-
-    return uniqueAssetName(resolved);
-  }
 
   function decorationText(tag: string, guid: string, mod?: string) {
     const tagInfo = allowedTags[tag];
@@ -54,31 +43,21 @@ export function activate(context: vscode.ExtensionContext) {
   }
 
   function updateDecorations() {
-    if (!activeEditor) {
-      return;
-    }
-
-    if (!minimatch(activeEditor.document.fileName, ASSETS_FILENAME_PATTERN, { dot: true })) {
+    if (!activeEditor || !editorFormats.isAnnoXml(activeEditor.document)) {
       return;
     }
 
     const modName = path.basename(utils.searchModPath(activeEditor.document.uri.fsPath));
 
-    const traverse = (activeEditor: vscode.TextEditor, color: string, 
-      regex: RegExp, onMatch: (match: RegExpExecArray) => string, 
+    const traverse = (activeEditor: vscode.TextEditor, color: string,
+      regex: RegExp, onMatch: (match: RegExpExecArray) => string,
       type: vscode.TextEditorDecorationType) => {
 
-      if (activeEditor.document.lineCount > 30000) {
-        // ignore 30k+ lines
+      if (!editorFormats.isAnnoXml(activeEditor.document)) {
         return;
       }
 
       const text = activeEditor.document.getText();
-      if (text.length > 1024 * 1024 * 10) {
-        // ignore 10MB+ files
-        return;
-      }
-
       const guids: vscode.DecorationOptions[] = [];
       let match;
       while ((match = regex.exec(text))) {
@@ -106,58 +85,7 @@ export function activate(context: vscode.ExtensionContext) {
       activeEditor.setDecorations(type, guids);
     };
 
-    const traverse2 = (activeEditor: vscode.TextEditor, color: string, 
-      regex: RegExp, onMatch: (match: RegExpExecArray) => string, 
-      type: vscode.TextEditorDecorationType) => {
-
-      if (activeEditor.document.lineCount > 30000) {
-        // ignore 30k+ lines
-        return;
-      }
-
-      const text = activeEditor.document.getText();
-      if (text.length > 1024 * 1024 * 10) {
-        // ignore 10MB+ files
-        return;
-      }
-
-      const guids: vscode.DecorationOptions[] = [];
-      let match;
-
-      let matches: Array<RegExpExecArray> = [];
-      while (match = regex.exec(text)) {
-        matches = [match, ...matches];
-      }
-
-      for (const match of matches) {
-        const startPos = activeEditor.document.positionAt(match.index);
-
-        const lineAbove = Math.max(0, startPos.line);
-        const lastIndex = activeEditor.document.lineAt(lineAbove).text.length;
-
-        const endPos = new vscode.Position(lineAbove, lastIndex);
-
-        const text = onMatch(match);
-        if (!text) {
-          continue;
-        }
-
-        const decoration: vscode.DecorationOptions = {
-          range: new vscode.Range(endPos, endPos),
-          renderOptions: {
-            after: {
-              contentText: text,
-              color: new vscode.ThemeColor(color),
-              margin: '3px'
-            }
-          }
-        };
-        guids.push(decoration);
-      }
-
-      activeEditor.setDecorations(type, guids);
-    };
-
+    // Decorate GUID numbers with name, template and origin (if applicable)
     let withinStandard = false;
     traverse(activeEditor, 'editorCodeLens.foreground', /(<\/?Standard>)|<(\w+)>(\d+)<\/\2>/g, (match) => {
       if (match[1] === '<Standard>') {
@@ -173,10 +101,7 @@ export function activate(context: vscode.ExtensionContext) {
       return '';
     }, guidDecorationType);
 
-    // traverse2(activeEditor, 'editorCodeLens.foreground', /(?<=GUID='|"|,| )(\d+)(?= |,|"|')/g, (match) => {
-    //   return shortGuidDecorator(match[1]);
-    // }, modOpDecorationType);
-
+    // Decorate top-level <Asset>s with name and template
     traverse(activeEditor, 'editorCodeLens.foreground', /<Asset>/g, (match) => {
       const doc = activeEditor?.document;
       if (!doc) {
@@ -220,34 +145,39 @@ export function activate(context: vscode.ExtensionContext) {
       timeout = undefined;
     }
     if (throttle) {
-      timeout = setTimeout(performance ? updateAssetAndPerformanceDecorations : updateDecorations, 500);
-    } else {
-      updateAssetAndPerformanceDecorations();
+      timeout = setTimeout(performance ? updateAssetAndPerformanceDecorations : updateDecorations,
+        2000 /* ms */);
+    } else if (performance) {
+      timeout = setTimeout(updateAssetAndPerformanceDecorations,
+        100 /* ms */);
+    }
+    else {
+      updateDecorations();
     }
   }
 
   if (activeEditor) {
-    triggerUpdateDecorations();
+    triggerUpdateDecorations(true, true);
   }
 
   vscode.window.onDidChangeActiveTextEditor(editor => {
     activeEditor = editor;
     if (editor) {
-      triggerUpdateDecorations();
-    }
-  }, null, context.subscriptions);
-
-  vscode.workspace.onDidChangeTextDocument(event => {
-    if (activeEditor && event.document === activeEditor.document) {
-      triggerUpdateDecorations(true);
       clearPerformanceDecorations();
+      triggerUpdateDecorations(false, true);
     }
   }, null, context.subscriptions);
 
+  vscode.workspace.onDidChangeTextDocument(async (event) => {
+    if (activeEditor && event.document === activeEditor.document) {
+      clearPerformanceDecorations();
+      triggerUpdateDecorations(true, false);
+    }
+  }, null, context.subscriptions);
 
-  vscode.workspace.onDidSaveTextDocument(event => {
+  vscode.workspace.onDidSaveTextDocument(async (event) => {
     if (activeEditor) {
-      triggerUpdateDecorations(true, true);
+      triggerUpdateDecorations(false, true);
     }
   }, null, context.subscriptions);
 }